requirements.txt
python-dotenv==0.15.0
discord.py==1.5.1
mcstatus==4.1.0